Add primaryNav to redux store and optimize NavBar code
