+++ src/redux/siteData/siteDataActions.js
@@ -1,4 +1,8 @@
-import { SET_COLOR_SCHEME, SET_PRIMARY_NAV } from './siteDataTypes';
+import {
+  SET_COLOR_SCHEME,
+  SET_PRIMARY_NAV,
+  SET_LOGIN_STATUS,
+} from './siteDataTypes';
 
 export const setPrimaryNav = userInput => ({
   type: SET_PRIMARY_NAV,
@@ -8,4 +12,9 @@ export const setPrimaryNav = userInput => ({
 export const setColorScheme = color => ({
   type: SET_COLOR_SCHEME,
   color,
+});
+
+export const setLoginStatus = loginStatus => ({
+  type: SET_LOGIN_STATUS,
+  loginStatus,
 });
