+++ src/redux/siteData/siteDataReducer.js
@@ -1,8 +1,9 @@
-import { SET_COLOR_SCHEME, SET_PRIMARY_NAV } from './siteDataTypes';
+import { SET_COLOR_SCHEME, SET_LOGIN_STATUS, SET_PRIMARY_NAV } from './siteDataTypes';
 
 const initialState = {
   colorScheme: 'dark',
   primaryNav: true,
+  loginStatus: false,
 };
 
 const reducer = (state = initialState, action) => {
@@ -16,6 +17,11 @@ const reducer = (state = initialState, action) => {
       return {
         ...state,
         colorScheme: action.color,
+      };
+    case SET_LOGIN_STATUS:
+      return {
+        ...state,
+        loginStatus: action.loginStatus,
       };
 
     default:
